HHH-17377 - Migrate to JPA 3.2
https://hibernate.atlassian.net/browse/HHH-17377

Test compilation success!!  95 test failures next

+++ hibernate-core/src/test/java/org/hibernate/orm/test/cut/CompositeUserTypeTest.java
@@ -109,7 +109,7 @@ public class CompositeUserTypeTest {
 					session.persist( f );
 					session.flush();
 					// Test value conversion during insert
-					BigDecimal amountViaSql = session.createNativeQuery(
+					BigDecimal amountViaSql = (BigDecimal) session.createNativeQuery(
 									"select amount_millions from MutualFund", BigDecimal.class )
 							.uniqueResult();
 					assertEquals( AMOUNT_MILLIONS.doubleValue(), amountViaSql.doubleValue(), 0.01d );
